Move column rendering tests
