+++ dbml/src/main/kotlin/com/zynger/floorplan/dbml/Column.kt
@@ -1,7 +1,6 @@
 package com.zynger.floorplan.dbml
 
 data class Column(
-    val rawValue: String,
     val name: String,
     val type: String,
     val note: String? = null,
@@ -9,7 +8,8 @@ data class Column(
     val primaryKey: Boolean = false,
     val notNull: Boolean = false,
     val increment: Boolean = false,
-    val reference: Reference? = null // not null when this column references another through a column attribute
+    val reference: Reference? = null, // not null when this column references another through a column attribute,
+    val rawValue: String = name
 ) {
     // example column: address varchar(255) [unique, not null, note: 'to include unit number']
 }
